clean up to individual methods
